Revert "update event to observable"
This reverts commit ef31c3a849ade40ffc5a88b11c2055f3031d5570.
