+++ examples/simple.html
@@ -7,35 +7,14 @@
 </head>
 
 <body>
-<div style="position: absolute;top: 1px;left: 2px;padding: 10px;color: #cccccc;background: rgba(0, 50, 50, 0.5);">
-  <button id="on" value="注册事件" onclick="on()">注册事件</button>
-  <button id="un" value="注销事件" onclick="un()">注销事件</button>
-  <button id="unAll" value="注销所有事件" onclick="unAll()">注销所有事件</button>
-  <textarea id="logtext"></textarea>
-</div>
-<div id="earth" style="width: 100%; height: 100%;"></div>
-<script type="application/javascript">
-  var earth = new fe.Earth("earth");
-  earth.addLayer({
-    type: 'rasterTile',
-    url: 'http://mt3.google.cn/vt/lyrs=s@138&hl=zh-CN&gl=CN&src=app&x={x}&y={y}&z={z}&s=Galil'
-  });
-  function on() {
-    earth.on('earth.roate',earthRoate);
-  }
-  var getEventsIds=0;
-  function earthRoate(e) {
-    getEventsIds++;
-    var textArea=document.getElementById('logtext');
-    textArea.value+=getEventsIds.toString()+":"+e.toString();
-  }
-  function un() {
-    earth.un('earth.roate',earthRoate);
-  }
-  function unAll() {
-    earth.unAll();
-  }
-</script>
+  <div id="earth" style="width: 100%; height: 100%;"></div>
+  <script type="application/javascript">
+    var earth = new fe.Earth("earth");
+    earth.addLayer({
+      type: 'rasterTile',
+      url: 'http://mt3.google.cn/vt/lyrs=s@138&hl=zh-CN&gl=CN&src=app&x={x}&y={y}&z={z}&s=Galil'
+    });
+  </script>
 </body>
 
 </html>
